requirements.txt
requests
python-dotenv
pydantic
py-clob-client-v2
py-builder-relayer-client
py-builder-signing-sdk
eth-abi
eth-utils
websocket-client
